+++ JianMerchant/src/main/java/com/woniukeji/jianmerchant/partjob/FilterItemClickListener.java
@@ -8,5 +8,5 @@ import android.view.View;
 public interface FilterItemClickListener {
     void onItemClick(int position);
 
-    boolean onItemLongClick(int position,View view,int login_id);
+    boolean onItemLongClick(int position,View view,long login_id);
 }
